requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
dropbox==12.0.2
easyocr==1.7.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
ply==3.11
pyclipper==1.3.0.post6
pyparsing==3.2.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
stone==3.3.1
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.0
torchvision==0.20.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
